feat: implement the starting game panel for horse betting, make basic game level scene configuration.
implement basic UI and its logic for betting on a specific horse by selecting it from a list.
implement adaptive layout of information about horses in the UI.
create some test data assets for horses (for testing).
configure level cameras and the post processing.
remove extra unnecessary colliders from imported props.
make some of the props in the level scene static.
add a temporary code to the GameManager to show the result in game mode (for testing).