feat: implement a visual horse indicator, add finish timing to the leaderboard, sync horse run speed with anim
import new horse icon.
implement a visual horse indicator (use the horse name + add a highlight to see player's horse)
implement the logic for determining the finish time and add it to the leaderboard information.
add new horses to the race.
recreate the prefab hierarchy for the horse for easy editing.
fix some little bugs with leaderboard and horses info cards:
 - inaccurate horse selection
 - occasional display of bottom places at the top of the leaderboard